use uint16_t for queue resource

+++ lib/src/mt_dev.c
@@ -344,7 +344,7 @@ int dev_rx_runtime_queue_start(struct mtl_main_impl* impl, enum mtl_port port) {
   int ret;
   struct mt_rx_queue* rx_queue;
 
-  for (int q = 0; q < inf->max_rx_queues; q++) {
+  for (uint16_t q = 0; q < inf->max_rx_queues; q++) {
     rx_queue = &inf->rx_queues[q];
     if (rx_queue->active) {
       ret = rte_eth_dev_rx_queue_start(inf->port_id, q);
@@ -1150,7 +1150,7 @@ int dev_reset_port(struct mtl_main_impl* impl, enum mtl_port port) {
 
   /* restore rte flow */
   struct mt_rx_queue* rx_queue;
-  for (uint16_t rx_q = 0; rx_q < (uint16_t)inf->max_rx_queues; rx_q++) {
+  for (uint16_t rx_q = 0; rx_q < inf->max_rx_queues; rx_q++) {
     rx_queue = &inf->rx_queues[rx_q];
     if (rx_queue->flow_rsp) {
       mt_rte_free(rx_queue->flow_rsp);
@@ -1425,7 +1425,7 @@ static int dev_if_uinit_rx_queues(struct mt_interface* inf) {
 
   if (!inf->rx_queues) return 0;
 
-  for (int q = 0; q < inf->max_rx_queues; q++) {
+  for (uint16_t q = 0; q < inf->max_rx_queues; q++) {
     rx_queue = &inf->rx_queues[q];
     if (rx_queue->active) {
       warn("%s(%d), rx queue %d still active\n", __func__, port, q);
@@ -1460,7 +1460,7 @@ static int dev_if_init_rx_queues(struct mtl_main_impl* impl, struct mt_interface
   }
 
   if (!mt_has_rx_mono_pool(impl)) {
-    for (uint16_t q = 0; q < (uint16_t)inf->max_rx_queues; q++) {
+    for (uint16_t q = 0; q < inf->max_rx_queues; q++) {
       rx_queues[q].queue_id = q;
       rx_queues[q].port = inf->port;
       rx_queues[q].port_id = inf->port_id;
@@ -1547,7 +1547,7 @@ static int dev_if_init_tx_queues(struct mtl_main_impl* impl, struct mt_interface
     return -ENOMEM;
   }
 
-  for (uint16_t q = 0; q < (uint16_t)inf->max_tx_queues; q++) {
+  for (uint16_t q = 0; q < inf->max_tx_queues; q++) {
     tx_queues[q].port = inf->port;
     tx_queues[q].port_id = inf->port_id;
     tx_queues[q].queue_id = q;
@@ -1770,7 +1770,7 @@ struct mt_rx_queue* mt_dev_get_rx_queue(struct mtl_main_impl* impl, enum mtl_por
   }
 
   mt_pthread_mutex_lock(&inf->rx_queues_mutex);
-  for (int q = 0; q < inf->max_rx_queues; q++) {
+  for (uint16_t q = 0; q < inf->max_rx_queues; q++) {
     rx_queue = &inf->rx_queues[q];
     if (rx_queue->active) continue;
     if (flow && flow->hdr_split) { /* continue if not hdr split queue */
